Fix for #1, clang errors on first build
The issue here was the actual definitions of some rake tasks depend on having gtest installed.  It was a little difficult to figure out how to work around this, but in the end, the fix wasn't too hard.  Just check for gtest, and if it isn't installed, avoid defining dependencies on the test objects.